Move most of the parsing logic to its own module
