+++ pytest_bdd/parser.py
@@ -1,0 +1,195 @@
+import io
+import os.path
+import re
+from collections import OrderedDict
+
+from pytest_bdd import types, exceptions
+from pytest_bdd.feature import Feature, Examples, get_tags, Scenario, Background, Step
+
+SPLIT_LINE_RE = re.compile(r"(?<!\\)\|")
+COMMENT_RE = re.compile(r"(^|(?<=\s))#")
+STEP_PREFIXES = [
+    ("Feature: ", types.FEATURE),
+    ("Scenario Outline: ", types.SCENARIO_OUTLINE),
+    ("Examples: Vertical", types.EXAMPLES_VERTICAL),
+    ("Examples:", types.EXAMPLES),
+    ("Scenario: ", types.SCENARIO),
+    ("Background:", types.BACKGROUND),
+    ("Given ", types.GIVEN),
+    ("When ", types.WHEN),
+    ("Then ", types.THEN),
+    ("@", types.TAG),
+    # Continuation of the previously mentioned step type
+    ("And ", None),
+    ("But ", None),
+]
+
+
+def split_line(line):
+    """Split the given Examples line.
+
+    :param str|unicode line: Feature file Examples line.
+
+    :return: List of strings.
+    """
+    return [cell.replace("\\|", "|").strip() for cell in SPLIT_LINE_RE.split(line[1:-1])]
+
+
+def parse_line(line):
+    """Parse step line to get the step prefix (Scenario, Given, When, Then or And) and the actual step name.
+
+    :param line: Line of the Feature file.
+
+    :return: `tuple` in form ("<prefix>", "<Line without the prefix>").
+    """
+    for prefix, _ in STEP_PREFIXES:
+        if line.startswith(prefix):
+            return prefix.strip(), line[len(prefix) :].strip()
+    return "", line
+
+
+def strip_comments(line):
+    """Remove comments.
+
+    :param str line: Line of the Feature file.
+
+    :return: Stripped line.
+    """
+    res = COMMENT_RE.search(line)
+    if res:
+        line = line[: res.start()]
+    return line.strip()
+
+
+def get_step_type(line):
+    """Detect step type by the beginning of the line.
+
+    :param str line: Line of the Feature file.
+
+    :return: SCENARIO, GIVEN, WHEN, THEN, or `None` if can't be detected.
+    """
+    for prefix, _type in STEP_PREFIXES:
+        if line.startswith(prefix):
+            return _type
+
+
+def parse_feature(basedir, filename, encoding="utf-8"):
+    """Parse the feature file.
+
+    :param str basedir: Feature files base directory.
+    :param str filename: Relative path to the feature file.
+    :param str encoding: Feature file encoding (utf-8 by default).
+    """
+    abs_filename = os.path.abspath(os.path.join(basedir, filename))
+    rel_filename = os.path.join(os.path.basename(basedir), filename)
+    self = Feature(
+        scenarios=OrderedDict(),
+        filename=abs_filename,
+        rel_filename=rel_filename,
+        line_number=1,
+        name=None,
+        tags=set(),
+        examples=Examples(),
+        background=None,
+        description="",
+    )
+    scenario = None
+    mode = None
+    prev_mode = None
+    description = []
+    step = None
+    multiline_step = False
+    prev_line = None
+
+    with io.open(abs_filename, "rt", encoding=encoding) as f:
+        content = f.read()
+
+    for line_number, line in enumerate(content.splitlines(), start=1):
+        unindented_line = line.lstrip()
+        line_indent = len(line) - len(unindented_line)
+        if step and (step.indent < line_indent or ((not unindented_line) and multiline_step)):
+            multiline_step = True
+            # multiline step, so just add line and continue
+            step.add_line(line)
+            continue
+        else:
+            step = None
+            multiline_step = False
+        stripped_line = line.strip()
+        clean_line = strip_comments(line)
+        if not clean_line and (not prev_mode or prev_mode not in types.FEATURE):
+            continue
+        mode = get_step_type(clean_line) or mode
+
+        allowed_prev_mode = (types.BACKGROUND, types.GIVEN, types.WHEN)
+
+        if not scenario and prev_mode not in allowed_prev_mode and mode in types.STEP_TYPES:
+            raise exceptions.FeatureError(
+                "Step definition outside of a Scenario or a Background", line_number, clean_line, filename
+            )
+
+        if mode == types.FEATURE:
+            if prev_mode is None or prev_mode == types.TAG:
+                _, self.name = parse_line(clean_line)
+                self.line_number = line_number
+                self.tags = get_tags(prev_line)
+            elif prev_mode == types.FEATURE:
+                description.append(clean_line)
+            else:
+                raise exceptions.FeatureError(
+                    "Multiple features are not allowed in a single feature file",
+                    line_number,
+                    clean_line,
+                    filename,
+                )
+
+        prev_mode = mode
+
+        # Remove Feature, Given, When, Then, And
+        keyword, parsed_line = parse_line(clean_line)
+        if mode in [types.SCENARIO, types.SCENARIO_OUTLINE]:
+            tags = get_tags(prev_line)
+            self.scenarios[parsed_line] = scenario = Scenario(self, parsed_line, line_number, tags=tags)
+        elif mode == types.BACKGROUND:
+            self.background = Background(feature=self, line_number=line_number)
+        elif mode == types.EXAMPLES:
+            mode = types.EXAMPLES_HEADERS
+            (scenario or self).examples.line_number = line_number
+        elif mode == types.EXAMPLES_VERTICAL:
+            mode = types.EXAMPLE_LINE_VERTICAL
+            (scenario or self).examples.line_number = line_number
+        elif mode == types.EXAMPLES_HEADERS:
+            (scenario or self).examples.set_param_names([l for l in split_line(parsed_line) if l])
+            mode = types.EXAMPLE_LINE
+        elif mode == types.EXAMPLE_LINE:
+            (scenario or self).examples.add_example([l for l in split_line(stripped_line)])
+        elif mode == types.EXAMPLE_LINE_VERTICAL:
+            param_line_parts = [l for l in split_line(stripped_line)]
+            try:
+                (scenario or self).examples.add_example_row(param_line_parts[0], param_line_parts[1:])
+            except exceptions.ExamplesNotValidError as exc:
+                if scenario:
+                    raise exceptions.FeatureError(
+                        """Scenario has not valid examples. {0}""".format(exc.args[0]),
+                        line_number,
+                        clean_line,
+                        filename,
+                    )
+                else:
+                    raise exceptions.FeatureError(
+                        """Feature has not valid examples. {0}""".format(exc.args[0]),
+                        line_number,
+                        clean_line,
+                        filename,
+                    )
+        elif mode and mode not in (types.FEATURE, types.TAG):
+            step = Step(name=parsed_line, type=mode, indent=line_indent, line_number=line_number, keyword=keyword)
+            if self.background and not scenario:
+                target = self.background
+            else:
+                target = scenario
+            target.add_step(step)
+        prev_line = clean_line
+
+    self.description = u"\n".join(description).strip()
+    return self
